fix: support email queue
