+++ website/.env.example
@@ -15,8 +15,10 @@
 # RESEND_API_KEY=re_...
 # RESEND_FROM_EMAIL="Docs Cloud <docs@notifications.example.com>"
 # ENTERPRISE_SUPPORT_TO_EMAIL=support@example.com
-# Optional: protects the enterprise support email worker when called outside Vercel Cron.
+# Required for the enterprise support email retry worker.
 # CRON_SECRET=...
+# Optional: use a different secret for the enterprise support email worker.
+# ENTERPRISE_SUPPORT_EMAIL_WORKER_SECRET=...
 
 # Optional: switch docs search to Algolia for the website.
 # DOCS_SEARCH_PROVIDER=algolia
